Moved spells into magic. Reporting errors.
