+++ src/index.css
@@ -25,3 +25,18 @@ h1, h2, h3, h4 {
 div {
 	position: relative;
 }
+
+
+/* ----- */
+
+.text-big {
+	font-size: 1.2em;
+}
+
+.text-mid {
+	font-size: 0.9em;
+}
+
+.text-small {
+	font-size: 0.8em;
+}
